Make alias autoconfiguration work in Bash
Bash hides its command history from other processes until it exits,
unless the Bash builtin "history -a" is executed. The builtin only
writes the history of the Bash process it has been called in, so there
is no way to programmatically call it from a Python program.

Due to this, the only way to get autoconfiguration to work is to disable
the last history item check for Bash.