extract constants to const file, separate logs.confirm_text function, remove confirmation log function, add env variable for double confirms
Moved the constants declared in the fix_command file to the dedicated constants file.

seperated the command_text file into two different functions so that there isn't another check for if its in the double confirm dict

moved double confirm logic from fix_command into select_command in the ui. I think this would make more sense because the underlying program shouldn't really care about confirmations.

making double confirm off by default with a environment variable required to turn it on. This might be helpful if we try to make a pr for this in the main repo.

update readme to reflect this