Added basic support for Atuin Atuin is shell history on steriods and is commonly used as an enhancement of the existing shell history. Unlike Bash or Zsh, Atuin uses an SQLite database to store the command history.
The nice thing about this however is that this makes Atuin shell
agnostic, so all we've done is add a basic check in the generic shell
class to check to see if the user provided a path to the Atuin database.
If so, then we check the database instead of the default location.

Right now this system has an issue where it'll reconnect to the database
multiple times, which may cause performance issues. The point however is
that this works, and that we can search the Atuin database.